BACKEND: moving prisma and updating /transfer to filter according to price and team and maxPrice
